mathify/requirements.txt
kivy==2.3.0
kivymd==1.1.1
pillow
pytesseract
numpy
sympy